pyproject.toml
[tool.poetry]
name = "sqlite-compactor"
version = "0.1.0"
description = "Dump rows from sqlite tables into parquet files."
authors = ["Nicolas Bonnet <mail@nicolasbon.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.2" #
python-ulid = "2.2.0"
duckdb = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
